PDFBOX-2576: use long values for the object number and integer values for the generation number
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1663130 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdfparser/BaseParser.java
@@ -48,8 +48,7 @@ import org.apache.pdfbox.cos.COSObjectKey;
  * This class is used to contain parsing logic that will be used by both the
  * PDFParser and the COSStreamParser.
  *
- * @author <a href="mailto:ben@benlitchfield.com">Ben Litchfield</a>
- * @version $Revision$
+ * @author Ben Litchfield
  */
 public abstract class BaseParser implements Closeable
 {
@@ -248,7 +247,7 @@ public abstract class BaseParser implements Closeable
             {
                 throw new IOException("expected number, actual=" + number + " at offset " + genOffset);
             }
-            COSObjectKey key = new COSObjectKey(((COSInteger) number).intValue(),
+            COSObjectKey key = new COSObjectKey(((COSInteger) number).longValue(),
                     ((COSInteger) generationNumber).intValue());
             retval = document.getObjectFromPool(key);
         }
@@ -994,7 +993,7 @@ public abstract class BaseParser implements Closeable
                     if (po.get(po.size()-1) instanceof COSInteger)
                     {
                         COSInteger number = (COSInteger)po.remove( po.size() -1 );
-                        COSObjectKey key = new COSObjectKey(number.intValue(), genNumber.intValue());
+                        COSObjectKey key = new COSObjectKey(number.longValue(), genNumber.intValue());
                         pbo = document.getObjectFromPool(key);
                     }
                     else
@@ -1658,9 +1657,9 @@ public abstract class BaseParser implements Closeable
      * @return the object number being read.
      * @throws IOException if an I/O error occurs
      */
-    protected int readObjectNumber() throws IOException
+    protected long readObjectNumber() throws IOException
     {
-        int retval = readInt();
+        long retval = readLong();
         if (retval < 0 || retval >= OBJECT_NUMBER_THRESHOLD)
         {
             throw new IOException("Object Number '" + retval + "' has more than 10 digits or is negative");
@@ -1826,7 +1825,7 @@ public abstract class BaseParser implements Closeable
     protected COSObjectKey parseObjectKey(boolean continueOnError) throws IOException
     {
         //we are going to parse a normal object
-        int number = -1;
+        long number = -1;
         int genNum;
         boolean missingObjectNumber = false;
         try
